Fixed bug that prevented the module from running on AWS lambda servers.
The pipe now uses C char* string operations to implement interprocess
communication. This is done because using std::string operations
caused a dependency on GLIBCXX_3.4.21, which does not exist on AWS
servers. There may be another way to achieve the same goal without
using `risky' operation, but I have yet to find it.